Adding Day1 solutions
I've finished the Day 1 exercise for the Advent Of Code 2020.
I approached the first part in two different ways. I wasn't sure
if a for loop would be faster than searching through the list so
I created a function for each. I added a little performance check
to compare the times.

I was surprised to see that the for loop ran 2x as fast as the
search. I wasn't expecting it to be that dramatic.

For part 2, I didn't create a new search() function since I felt
it wasn't a good approach. As such, I only made a for loop function
to figure out the winning combo. I've attached the output log in the
Day1/ directory.