Adding Day 2 part 1 solution
I've completed the first part of Day 2's password validation
challenge. I used regular expressions to help me parse through all
of the values. I broke apart the password contraints and compiled it
into a regular expression and then tried to match it against the
password. If it matched, then I would save the combo and add it
to the valid password list.