Adding Day2 part 2 code
I added the second part of Day 2's exercise, which was surprisingly straightforward. I was expecting it to be a little bit more tricky, but all I needed to do was search through the password to find the characters at the specified position.

I've included the output into the Day2 directory.