Adding Day 3 code (#3)
* Adding template.py file

Adding a template python file so that starting up a new
exercise is smoother. The file handles the basic input
and file reading.

* Adding Day 3 code

I've completed Day 3's exercise in traversal. I thought it was
a fun little exercise navigating through the 2d list. I had
made the slope configurable for the first part since it seemed
like a smart way to approach it, which made the second part
very easy. All I had to do was pass in the right slope
and it would output the number of trees hit.

For convenience sake, I hard-coded the question's slopes into
the main function. But it would be easy to revert that and
update the program to read in a slope or list of slopes.