Adding Day 4 solution (#4)
* Adding Day 4 part 2

I didn't have too much trouble with the second part of Day 4.
It seemed to be a good regex refresher for various formats.
I had a small issue where I misread the requirement for the
byr field, but after that, it worked as intended.

* Adding Day 4 log files and removing *.log from gitignore

Adding the log files for Day 4. They had been excluded due
to the gitignore ignoring *.log files. I have removed that since
I feel like .log files can be useful. But if they get too troublesome,
I can re-add it to the gitignore