Adding Day 7 part 2 solution
I had some trouble getting the recursion for this part working.
My answers were always just a little too small. I needed to add
some additional logging as well as a few extra test input files
to figure out what was going on. In the end, I needed to set
the base total to 1 instead of 0 to account for the parent bag.

It was a good exercise in creating my own tests as well as
a good refresher on recursion.