Adding Day 8 part 2
I tried to approach this using a linked list since that
seemed like a good way to refresh my memory on the concept.
Unfortunately, that led to a lot of issues down the line
as I hit recursion limits. I probably could have handled it
better, but I decided to drop it and reuse my function from
part 1. That was probably the intended method, but I wanted
to try something fun.

The linked list way worked on small inputs, but broke when
the inputs started to get large. It was sad to see, but at least
I got plenty of experience with linked lists. I left the code inside
the accumulator.py file.