Updating template file format
I'm updating the template python file so that
it uses a class-based structure. I feel like this
style will be better since I can store the input
in a class variable instead of having to pass around
the input list to every function. Hopefully this is
a useful change and makes it easier in the future.

Also removing some leftover comments from previous python
files.