Adding Day 10 solutions
Part 1 was super easy to solve. I had to read the
question a couple of times before I fully undrestood it.
Part 2, however, was a different beast. I tried a few
methods to solve it, but nothing panned out. I needed to look
up some other approaches to the solution. In my search I
came across Dynamic Programming. This was a helpful concept
that allowed me to get the right answer, but I wish I was able
to come up with it on my own.

But, I suppose that is the point of doing these exercises - they
are a great way to learn more.