Repurposing and restructuring
I am going to repurpose this repository to only
handle the Advent of Code 2020 programs. This will
be much better instead of having a general repository
for all of my projects. Since I am repurposing this
repo, I am also restructuring it to remove the
AdventOfCode2020 directory. Everything should
go up a level, but other than that, no major changes.