Adding Day 12 part 2
The second part of this challenge was pretty
easy, just like the first part. It took a while
to conceptualize the waypoint rotation, but once
I drew it out, it made it pretty easy. I needed to
figure out how to modify my existing codebase so that
it could handle both the waypoint and non-waypoint
navigation. Once I got that sorted out, it was
pretty straightforward.