Adding Day 13 solution
Part 1 was pretty easy, but part 2 was a doozy.
I misunderstood the problem multiple times which
caused me to approach it wrong. Once I finally grasped
the concept, I tried to solve it by brute forcing it. It
worked for all of the sample inputs, but it
obviously didn't work against the large input set. I
created a fun peek() function to help me debug. It would create
a fun little graphic to show which buses were leaving when.

I needed to look up some help for part 2. I came across
Chinese Remainder theorem which helped me solve it.
I looked up some videos on the theorem and I understand
it somewhat. I will need to take another look at it later
to make sure it sinks in.