Adding Day 14 part 1 solutions
I had a pretty easy time with this problem. I tried
searching for a while to figure out a good way to
handle the bit masking, but nothing seemed to do what
I wanted. I eventually decided to just replace it
character by character, but I still feel like there
should be a better way to handle it.