More progress towards Day 14
I had a lot of trouble on this day solely because
I keep misinterpreting the question. I have reworked
my solution so many times, but I think I'm finally
on the right track. I currently have an issue with the
X ->0/1 replacement, but once I get that fixed, then
the program should be working