Initial work on day 15
Adding my skeleton for day 15 part 1. I didn't have much
time to work on it, but I'm starting to figure out how
to get it working. The problem seems weird and hard
to explain so I'm not looking forward to the amount of
times I will need to reread it.